created syncuids.go with code from utils.go
